// front-upc-fertismart/src/app/shared/enums/endpoints.enum.ts
export enum Endpoints {
    SEGCEN_USERS = '/segcen/api/v1/users/',
    SEGCEN_USER_ACTIVATE = '/segcen/api/v1/users/activate/',
    SEGCEN_USER_DEACTIVATE = '/segcen/api/v1/users/deactivate/',
    SEGCEN_USER_REST_PASS = '/segcen/api/v1/users/reset_password/',
    SEGCEN_GROUPS = '/segcen/api/v1/groups/',
    ENTICEN_MANY_PARAM = '/api/many_parameters/',
    ENTICEN_PARAMETER = '/enticen/api/v1/parameter/',
    ENTICEN_PARAMETER_DETAIL = '/enticen/api/v1/parameter_detail/',
    ENTICEN_PARAMETER_DETAIL_LIST = '/enticen/api/v1/parameter_detail/detail_list/',
    FERTISMART_CULTIVO = '/fertismart/api/v1/cultivo/',
    FERTISMART_FERTILIZANTE = '/fertismart/api/v1/fertilizante/',
    FERTISMART_SUELO = '/fertismart/api/v1/tipo-suelo/',
    FERTISMART_VARIEDAD = '/fertismart/api/v1/variedad-cultivo/',
    FERTISMART_CAMPO = '/fertismart/api/v1/campo/',
    FERTISMART_APLICACION = '/fertismart/api/v1/historia-aplicacion/',
    FERTISMART_TIPO_GASTO = '/fertismart/api/v1/tipo-gasto/',
    FERTISMART_GASTO = '/fertismart/api/v1/gasto/',
    FERTISMART_FAQ = '/fertismart/api/v1/question-answer/',
    FERTISMART_RECOMENDACION = '/fertismart/api/v1/recomendacion/',
    FERTISMART_FINDER_DNI = '/fertismart/api/v1/get_person_dni/',
    FERTISMART_FINDER_RUC = '/fertismart/api/v1/get_person_ruc/',
    FERTISMART_DASHBOARD = '/fertismart/api/v1/dashboard/',
}